changes to support building on MacOS 10.9 and Xcode 6 (clang)
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@69337 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ centutils/libcentutils/ctstrutils.c
@@ -395,7 +395,7 @@ CTAllocateStringPrintfV(
     char *smallBuffer;
     unsigned int bufsize;
     int requiredLength;
-    unsigned int newRequiredLength;
+    ssize_t newRequiredLength;
     PSTR outputString = NULL;
     va_list args2;
