+++ lwadvapi/unthreaded/memory/lwstr.c
@@ -210,7 +210,7 @@ LwAllocateStringPrintfV(
     PSTR  pszSmallBuffer = NULL;
     DWORD dwBufsize = 0;
     INT   requiredLength = 0;
-    DWORD dwNewRequiredLength = 0;
+    ssize_t dwNewRequiredLength = 0;
     PSTR  pszOutputString = NULL;
     va_list args2;
